Show current key (no auto update)

+++ RDOTool.py
@@ -5,16 +5,7 @@ from tkinter import filedialog
 
 win=tk.Tk()
 win.title('RDO Tool')
-win.geometry('600x300')
-
-uniqueKey = tk.StringVar()
-uniqueKey_label = tk.Label(win, text="Unique key: ")
-uniqueKey_label.pack(fill='x', expand=True)
-uniqueKey_label.place(x = 210, y = 70)
-
-uniqueKey_entry = tk.Entry(win, textvariable=uniqueKey)
-uniqueKey_entry.pack(fill='x', expand=True)
-uniqueKey_entry.place(x = 210, y = 90)
+win.geometry('700x300')
 
 def choosePath():
     root = tk.Tk()
@@ -23,6 +14,7 @@ def choosePath():
     while file_path == "":
         messagebox.showerror("Choose Folder", "No Path Selected !")
         file_path = filedialog.askdirectory()
+        root.update()
         showPath()
     else:
         file = open("path.txt", "w")
@@ -32,19 +24,19 @@ def choosePath():
         showPath()
 
 def showPath():
-    # try:
+    if (os.path.exists("path.txt")):
         path = open("path.txt", "r")
         value = path.readline()
         path.close()
         currentPath = tk.Label(win, text="Current Folder Path:")
         currentPath.pack(fill='x', expand=True)
-        currentPath.place(x = 210, y=30)
+        currentPath.place(x = 300, y=30)
         
         currentPathValue = tk.Label(win, text=value)
         currentPathValue.pack(fill='x', expand=True)
-        currentPathValue.place(x = 210, y=50)
-    # except FileNotFoundError:
-        # choosePath()
+        currentPathValue.place(x = 300, y=50)
+    else:
+        pass
 
 def checkACL():
     path = open("path.txt", "r")
@@ -59,28 +51,36 @@ def checkACL():
         messagebox.showerror("RDO Tool", "Restart the program with Administrator permission to access this folder !")
 
 def installFile():
-    path = open("path.txt", "r")
-    value = path.readline()
-    path.close()
+    try:
+        path = open("path.txt", "r")
+        value = path.readline()
+        path.close()
     
-    bootLauncherFlowYMT = value + '/x64/' + 'boot_launcher_flow.ymt'
-    startupMETA = value + '/x64/data/' + 'startup.meta'
-    try: 
-        urlBootLauncherFlowYMT = "https://raw.githubusercontent.com/TheTimeLord32/Red-Dead-Online-Tool/master/src/boot_launcher_flow.ymt"
-        r1 = requests.get(urlBootLauncherFlowYMT)
-        with open(bootLauncherFlowYMT, 'wb') as f:
-            f.write(r1.content)
-        f.close()
-    
-        urlstartupMETA = "https://raw.githubusercontent.com/TheTimeLord32/Red-Dead-Online-Tool/master/src/startup.meta"
-        r2 = requests.get(urlstartupMETA)
-        with open(startupMETA, 'wb') as f:
-            f.write(r2.content)
-        f.close()
+        bootLauncherFlowYMT = value + '/x64/' + 'boot_launcher_flow.ymt'
+        startupMETA = value + '/x64/data/' + 'startup.meta'
+
+        if (os.path.isfile(bootLauncherFlowYMT) | os.path.isfile(startupMETA)):
+            messagebox.showinfo("Install file", "Files already installed !")
+            pass
+        else:
+            try: 
+                urlBootLauncherFlowYMT = "https://raw.githubusercontent.com/TheTimeLord32/Red-Dead-Online-Tool/master/src/boot_launcher_flow.ymt"
+                r1 = requests.get(urlBootLauncherFlowYMT)
+                with open(bootLauncherFlowYMT, 'wb') as f:
+                    f.write(r1.content)
+                f.close()
+            
+                urlstartupMETA = "https://raw.githubusercontent.com/TheTimeLord32/Red-Dead-Online-Tool/master/src/startup.meta"
+                r2 = requests.get(urlstartupMETA)
+                with open(startupMETA, 'wb') as f:
+                    f.write(r2.content)
+                f.close()
+            except FileNotFoundError:
+                messagebox.showerror("Install file", "Needed folders not found. Rechoose Path !")
+            else:
+                messagebox.showinfo("Install file", "Files installed successfully !")
     except FileNotFoundError:
-        messagebox.showerror("Install file", "Needed folders not found. Rechoose Path !")
-    else:
-        messagebox.showinfo("Install file", "Files installed successfully !")
+            messagebox.showerror("Install file", "Path not found. Choose folder first")
 
 def readFile():
     path = open("path.txt", "r")
@@ -128,10 +128,11 @@ def findOldString():
                 secondline = file_contents[1].strip()
 
         oldKey = secondline[4: -3]
+        win.update()
         return oldKey
 
     except FileNotFoundError:
-        messagebox.showerror("Files not found!")
+        return "No files yet installed"
 
 def replaceString():
     try:
@@ -141,36 +142,57 @@ def replaceString():
         boot_enable = directory + "/x64/boot_launcher_flow.ymt"
         startup_enable = directory + "/x64/data/startup.meta"
         old_string = findOldString()
-        new_string = uniqueKey.get()
+        new_string = newUniqueKey.get()
 
-        files = [boot_enable, startup_enable]
-        for file_path in files:
-            # Read the contents of the file
-            with open(file_path, 'r') as file:
-                file_contents = file.read()
+        if (not new_string):
+            messagebox.showerror("Replace unique key", "Empty unique key")
+        else:
+            files = [boot_enable, startup_enable]
+            for file_path in files:
+                # Read the contents of the file
+                with open(file_path, 'r') as file:
+                    file_contents = file.read()
 
-            # Replace the old string with the new string
-            modified_contents = file_contents.replace(old_string, new_string)
+                # Replace the old string with the new string
+                modified_contents = file_contents.replace(old_string, new_string)
 
-            # Write the modified contents back to the file
-            with open(file_path, 'w') as file:
-                file.write(modified_contents)
+                # Write the modified contents back to the file
+                with open(file_path, 'w') as file:
+                    file.write(modified_contents)
 
-        messagebox.showinfo("Replace unique key", "Unique key saved successfully!")
+            messagebox.showinfo("Replace unique key", "Unique key saved successfully!")
     except FileNotFoundError:
         messagebox.showerror("Replace unique key", "Path not found!")
 
-folder = tk.Button(win, text="Choose Folder", width=10, height=5, command=choosePath)
-install = tk.Button(win, text="Install file", width=10, height=5, command=installFile)
-activate = tk.Button(win, text = "Activate file", width=10, height=5, command=activateFile)
-deactivate = tk.Button(win, text = "Deactivate file", width=10, height=5, command=deactivateFile)
-replace = tk.Button(win, text="Replace unique key", width=25, height=5, command=replaceString)
+folder = tk.Button(win, text="Choose Folder", width=15, height=5, command=choosePath)
+install = tk.Button(win, text="Install file", width=15, height=5, command=installFile)
+activate = tk.Button(win, text = "Activate file", width=15, height=5, command=activateFile)
+deactivate = tk.Button(win, text = "Deactivate file", width=15, height=5, command=deactivateFile)
+replace = tk.Button(win, text="Replace unique key", width=32, height=5, command=replaceString)
+
+showPath()
+currentUniqueKey_label = tk.Label(win, text="Current unique key: ")
+currentUniqueKey_label.pack(fill='x', expand=True)
+currentUniqueKey_label.place(x = 300, y = 70)
+
+currentUniqueKeyValue = tk.Label(win, text=findOldString())
+currentUniqueKeyValue.pack(fill='x', expand=True)
+currentUniqueKeyValue.place(x = 425, y = 70)
+
+newUniqueKey = tk.StringVar()
+newUniqueKey_label = tk.Label(win, text="New unique key: ")
+newUniqueKey_label.pack(fill='x', expand=True)
+newUniqueKey_label.place(x = 300, y = 90)
+
+newUniqueKey_entry = tk.Entry(win, textvariable=newUniqueKey)
+newUniqueKey_entry.pack(fill='x', expand=True)
+newUniqueKey_entry.place(x = 425, y = 90)
 
 folder.place(x = 30,y = 30)
-install.place(x = 120, y = 30)
+install.place(x = 150, y = 30)
 activate.place(x = 30, y = 120)
-deactivate.place(x = 120, y = 120)
-replace.place(x = 210, y = 120)
+deactivate.place(x = 150, y = 120)
+replace.place(x = 30, y = 210)
 
 win.mainloop()
 os._exit(0)
